Use separate component for balance string
